Save 5.3 Enemy Save Quase completo

+++ Assets/Script/Save/SaveController.cs
@@ -3,13 +3,19 @@ using System.IO;
 using Unity.Cinemachine;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 
 public class SaveController : MonoBehaviour
 {
     private string _saveLocation;
 
+    public List<GameObject> _enemysInPlataform = new List<GameObject>();
+    public List<GameObject> _enemysInGame = new List<GameObject>();
+
     void Start()
     {
+        Resources.FindObjectsOfTypeAll<Monster>().ToList().ForEach(m=>_enemysInGame.Add(m.gameObject));
+
         _saveLocation = Path.Combine(Application.persistentDataPath, "SaveData.json");
 
         LoadGame();
@@ -17,6 +23,26 @@ public class SaveController : MonoBehaviour
 
     public void SaveGame() 
     {
+        Player player = FindAnyObjectByType<Player>();
+        UIStatusManager statusManager = FindAnyObjectByType<UIStatusManager>();
+
+        if (player == null || statusManager == null)
+        {
+            return;
+        }
+
+        List<string> itemNames = new List<string>();
+        List<int> itemCount = new List<int>();
+
+        if (player.Inventory != null)
+        {
+            foreach (ItensData item in player.Inventory)
+            {
+                itemNames.Add(item.Name);
+                itemCount.Add(item.Count);
+            }
+        }
+
         SaveData saveData = new SaveData
         {
             _playerPosition = GameObject.FindGameObjectWithTag("Player").transform.position,
@@ -33,6 +59,9 @@ public class SaveController : MonoBehaviour
             _enemy = FindAnyObjectByType<UIStatusManager>().Enemy,
             _player = GameObject.FindGameObjectWithTag("Player").name,
 
+            _itensInventoryname = itemNames,
+            _itensInventorycount = itemCount,
+
         };
 
         File.WriteAllText(_saveLocation, JsonUtility.ToJson(saveData));
@@ -41,14 +70,16 @@ public class SaveController : MonoBehaviour
 
     public void SaveCombate()
     {
-        SaveData saveData = new SaveData
+        if (File.Exists(_saveLocation))
         {
-            _level = FindAnyObjectByType<Player>().Level,
-            _xp = FindAnyObjectByType<Player>().Xp,
-            _money = FindAnyObjectByType<Player>().Money,
-        };
+            SaveData saveData = JsonUtility.FromJson<SaveData>(File.ReadAllText(_saveLocation));
+            saveData._money = FindAnyObjectByType<Player>().Money;
+            saveData._xp = FindAnyObjectByType<Player>().Xp;
+            saveData._level = FindAnyObjectByType<Player>().Level;
 
-        File.WriteAllText(_saveLocation, JsonUtility.ToJson(saveData));
+            File.WriteAllText(_saveLocation, JsonUtility.ToJson(saveData));
+
+        }
     }
 
     public void LoadGame()
